refactor(frontend) reorganize page component forms
