+++ frontend/app/routes/hr-advisor/employee-profile/personal-information.tsx
@@ -17,8 +17,8 @@ import { BackLink } from '~/components/back-link';
 import { HttpStatusCodes } from '~/errors/http-status-codes';
 import { getTranslation } from '~/i18n-config.server';
 import { handle as parentHandle } from '~/routes/layout';
-import { PersonalInformationForm } from '~/routes/page-components/employees/personal-information/form';
-import { personalInformationSchema } from '~/routes/page-components/employees/validation.server';
+import { PersonalInformationForm } from '~/routes/page-components/profile/personal-information/form';
+import { personalInformationSchema } from '~/routes/page-components/profile/validation.server';
 import { toE164 } from '~/utils/phone-utils';
 import { formString } from '~/utils/string-utils';
