+++ frontend/app/routes/page-components/profile/referral-preferences/form.tsx
@@ -32,7 +32,7 @@ import type { InputRadiosProps } from '~/components/input-radios';
 import { InputSelect } from '~/components/input-select';
 import { REQUIRE_OPTIONS } from '~/domain/constants';
 import type { I18nRouteFile } from '~/i18n-routes';
-import type { Errors } from '~/routes/page-components/employees/validation.server';
+import type { Errors } from '~/routes/page-components/profile/validation.server';
 import { extractValidationKey } from '~/utils/validation-utils';
 
 interface ReferralPreferencesFormProps {
